[HRA-3]Merged master into feature/hra-3, almost final working login and registration sistem with dummy login test page
