+++ src/main/java/HRA/model/User.java
@@ -5,14 +5,22 @@ public class User {
     private String username;
     private String password;
     private String role;
+    private String name;
+    private String address;
+    private String email;
+    private String phoneNumber;
 
     public User() {
     }
 
-    public User(String username, String password, String role) {
+    public User(String username, String password, String role, String name, String address, String email, String phoneNumber) {
         this.username = username;
         this.password = password;
         this.role = role;
+        this.name=name;
+        this.address = address;
+        this.email = email;
+        this.phoneNumber = phoneNumber;
     }
 
     public String getUsername() {
@@ -38,6 +46,22 @@ public class User {
     public void setRole(String role) {
         this.role = role;
     }
+
+    public String getName() { return name; }
+
+    public void setName(String name) { this.name = name; }
+
+    public String getAddress() { return address; }
+
+    public void setAddress(String address) { this.address = address; }
+
+    public String getEmail() { return email; }
+
+    public void setEmail(String email) { this.email = email; }
+
+    public String getPhoneNumber() { return phoneNumber; }
+
+    public void setPhoneNumber(String phoneNumber) { this.phoneNumber = phoneNumber; }
 
     @Override
     public boolean equals(Object o) {
@@ -65,6 +89,10 @@ public class User {
                 "username='" + username + '\'' +
                 ", password='" + password + '\'' +
                 ", role='" + role + '\'' +
+                ", full name/hotel name='" + name + '\'' +
+                ", address='" + address + '\'' +
+                ", email='" + email + '\'' +
+                ", phone number='" + phoneNumber + '\'' +
                 '}';
     }
 }
