--- src/main/java/registration/Main.java
@@ -1,27 +1,0 @@
-package registration;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import registration.services.UserService;
-
-
-public class Main extends Application {
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage primaryStage) throws Exception {
-
-        UserService.loadUsersFromFile();
-
-        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("register.fxml"));
-        primaryStage.setTitle("Hotel Registration Application");
-        primaryStage.setScene(new Scene(root, 500, 450));
-        primaryStage.show();
-    }
-}
