--- src/main/java/registration/exceptions/CouldNotWriteUsersException.java
@@ -1,4 +1,0 @@
-package registration.exceptions;
-
-public class CouldNotWriteUsersException extends RuntimeException {
-}
