--- src/main/java/registration/model/User.java
@@ -1,101 +1,0 @@
-package registration.model;
-
-
-public class User {
-
-    private String username;
-    private String password;
-    private String role;
-    private String name;
-    private String address;
-    private String email;
-    private String phoneNumber;
-
-    public User() {
-    }
-
-    public User(String username, String password, String role, String name, String address, String email, String phoneNumber) {
-        this.username = username;
-        this.password = password;
-        this.role = role;
-        this.name=name;
-        this.address = address;
-        this.email = email;
-        this.phoneNumber = phoneNumber;
-
-
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-
-    public String getName() { return name; }
-
-    public void setName(String name) { this.name = name; }
-
-    public String getAddress() { return address; }
-
-    public void setAddress(String address) { this.address = address; }
-
-    public String getEmail() { return email; }
-
-    public void setEmail(String email) { this.email = email; }
-
-    public String getPhoneNumber() { return phoneNumber; }
-
-    public void setPhoneNumber(String phoneNumber) { this.phoneNumber = phoneNumber; }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        User user = (User) o;
-
-        if (!username.equals(user.username)) return false;
-        if (!password.equals(user.password)) return false;
-        return role.equals(user.role);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = username.hashCode();
-        result = 31 * result + password.hashCode();
-        result = 31 * result + role.hashCode();
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return "UserDTO{" +
-                "username='" + username + '\'' +
-                ", password='" + password + '\'' +
-                ", role='" + role + '\'' +
-                ", full name/hotel name='" + name + '\'' +
-                ", address='" + address + '\'' +
-                ", email='" + email + '\'' +
-                ", phone number='" + phoneNumber + '\'' +
-                '}';
-    }
-}
